+++ internal/feature/example/store/models.go
@@ -10,6 +10,12 @@ import (
 	"github.com/jackc/pgx/v5/pgtype"
 )
 
+type Cache struct {
+	Key       string
+	Value     []byte
+	ExpiresAt pgtype.Timestamptz
+}
+
 type Item struct {
 	ID        int64
 	Title     string